Changed reader lambda literal syntax
Changed from '\' followed by anything, to '#' followed by '('. Because the latter is already defined as a vector in regular scheme, most scheme IDEs have syntax highlighting for it (making for a better dev experience). Since we use '[]' to denote vectors in escheme, we have that syntax up for grabs for lambda literals instead! Inspired by clojure.