Enables more efficient method name-binding for <define-class> and <define-interface>
Now binds method names upon initial ctor of classes/interfaces, rather than only when bound to a symbol at runtime in the environment. Note that this will only change <define-class> and <define-interface> to be efficient, & that regular anonymous classes bound to symbols will have the exact same behavior as before.