Fixes a bug with <quote> and threads.
Quote used to just embed the given value directly into bytecode at compile-time. However, that meant that different threads executing the same function body would be accessing the same value in the bytecode of said function body, which for vectors and hashmaps meant that mutable state was accidentally being shared across threads!

We now get rid of the <load-symbol> instruction, and replace its combined use with the <load> instruction to instead have <load> be on its own, and having a new instruction called <quote> entirely handle the process of deep-copying mutable reader state at runtime, so as to ensure each thread's execution of the function body gets a newly allocated value when quoting a vector or hashmap!